% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Get features for training}
\usage{
get_features(con, date_from, date_to, features, trigger_name)
}
\arguments{
\item{con}{A feature connection as created by \code{connect_store()}.}

\item{date_from}{The startndate of the time window for training as a string.}

\item{date_to}{The end date of the time window for training as a string.}

\item{features}{A character vector of full feature names.}

\item{trigger_name}{A string with the name of the trigger, as defined in the feature's definition yaml.}
}
\value{
A data-frame.
}
\description{
Get features for training
}
\examples{
\dontrun{
training_data <- get_features(
  con = pargo_con,
  date_from = "2020-01-01",
  date_to = "2021-01-01",
  features = c("avg_order_amount_per_debtor_uuid", "n_complete_orders_per_email"),
  trigger_name = "order_uuid" # as defined in the feature's yaml definition
)
}
}
